# stream.hls/plugins.vimeo: update/fix