@bastimeyer's changes to `stream.hls` and my updates to `plugins.vimeo`.

closes #3492

I don't think this affects VODs; as far as I know the expiry times on them are sufficient to allow real-time streaming without exceeding the expiry.

I don't know if DASH livestreams are affected by this expiry issue, but I have not been able to find any to test with.  For the time being, this fixes the expiry issue with HLS and muxed-HLS streams.

Before HLS:
```
$ time streamlink -o /dev/null https://vimeo.com/718108412 240p
[cli][info] Found matching plugin vimeo for URL https://vimeo.com/718108412
[cli][info] Available streams: 1080p_hls, 720p_hls, 540p_hls, 360p_hls, 240p_hls, 240p (worst), 360p, 540p, 720p, 1080p (best)
[cli][info] Opening stream: 240p (hls)
[download][null] Written 176.1 MB (59m39s @ 53.1 KB/s)
[stream.hls][warning] Failed to reload playlist: Unable to open URL: https://live.vimeocdn.com/1654701977-0xc7beda8d65d85fe591f394cb296fc8110b4372b3/a6c4af14-cfcc-4aab-afba-d9372b08ec51/chunklist_b416000.m3u8 (410 Client Error: Gone for url: https://live.vimeocdn.com/1654701977-0xc7beda8d65d85fe591f394cb296fc8110b4372b3/a6c4af14-cfcc-4aab-afba-d9372b08ec51/chunklist_b416000.m3u8)
[ ... ]
[stream.hls][warning] Failed to reload playlist: Unable to open URL: https://live.vimeocdn.com/1654701977-0xc7beda8d65d85fe591f394cb296fc8110b4372b3/a6c4af14-cfcc-4aab-afba-d9372b08ec51/chunklist_b416000.m3u8 (410 Client Error: Gone for url: https://live.vimeocdn.com/1654701977-0xc7beda8d65d85fe591f394cb296fc8110b4372b3/a6c4af14-cfcc-4aab-afba-d9372b08ec51/chunklist_b416000.m3u8)
error: Error when reading from stream: Read timeout, exiting
[cli][info] Stream ended
[cli][info] Closing currently open stream...

real    60m42.844s
```
Before muxed-HLS:
```
$ time streamlink -o /dev/null https://vimeo.com/718108412 240p_hls
[cli][info] Found matching plugin vimeo for URL https://vimeo.com/718108412
[cli][info] Available streams: 1080p_hls, 720p_hls, 540p_hls, 360p_hls, 240p_hls, 240p (worst), 360p, 540p, 720p, 1080p (best)
[cli][info] Opening stream: 240p_hls (hls-multi)
[utils.named_pipe][info] Creating pipe streamlinkpipe-12688-1-1708
[utils.named_pipe][info] Creating pipe streamlinkpipe-12688-2-9617
[download][null] Written 179.3 MB (59m58s @ 44.1 KB/s)
[stream.hls][warning] Failed to reload playlist: Unable to open URL: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b360000_cmaf_v.m3u8 (403 Client Error: Forbidden for url: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b360000_cmaf_v.m3u8)
[stream.hls][warning] Failed to reload playlist: Unable to open URL: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b56000_cmaf_a.m3u8 (403 Client Error: Forbidden for url: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b56000_cmaf_a.m3u8)
[ ... ]
[stream.ffmpegmux][error] Pipe copy aborted: /tmp/streamlinkpipe-12688-2-9617
[ ... ]
[stream.hls][warning] Failed to reload playlist: Unable to open URL: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b360000_cmaf_v.m3u8 (403 Client Error: Forbidden for url: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b360000_cmaf_v.m3u8)
[stream.hls][warning] Failed to reload playlist: Unable to open URL: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b56000_cmaf_a.m3u8 (403 Client Error: Forbidden for url: https://llhls-live.akamaized.net/exp=1654702024~acl=%2Fa6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls%2F%2A~hmac=bce987bf4fb6f5d03c211d9a03d606e3cd68f5c278f098ca1db347360a9c6054/a6c4af14-cfcc-4aab-afba-d9372b08ec51_llhls/chunklist_b56000_cmaf_a.m3u8)
[stream.ffmpegmux][error] Pipe copy aborted: /tmp/streamlinkpipe-12688-1-1708
[download][null] Written 179.4 MB (1h1m58s @ 4.5 KB/s)
[cli][info] Stream ended
[cli][info] Closing currently open stream...

real    62m2.717s
```
After HLS (filtered):
```
$ time sl/bin/streamlink -l trace -o /dev/null https://vimeo.com/720451389 240p
[17:47:56.742511][cli][debug] OS:         Linux-4.9.0-18-amd64-x86_64-with-debian-9.13
[17:47:56.743242][cli][debug] Python:     3.7.3
[17:47:56.743651][cli][debug] Streamlink: 4.1.0+7.gc319aa44
[17:47:56.745891][cli][debug] Dependencies:
[17:47:56.752562][cli][debug]  isodate: 0.6.1
[17:47:56.754047][cli][debug]  lxml: 4.9.0
[17:47:56.756365][cli][debug]  pycountry: 22.3.5
[17:47:56.757504][cli][debug]  pycryptodome: 3.14.1
[17:47:56.758739][cli][debug]  PySocks: 1.7.1
[17:47:56.759913][cli][debug]  requests: 2.27.1
[17:47:56.761488][cli][debug]  websocket-client: 1.3.2
[17:47:56.763065][cli][debug]  importlib-metadata: 4.11.4
[17:47:56.764461][cli][debug] Arguments:
[17:47:56.764686][cli][debug]  url=https://vimeo.com/720451389
[17:47:56.764868][cli][debug]  stream=['240p']
[17:47:56.765081][cli][debug]  --loglevel=trace
[17:47:56.765301][cli][debug]  --output=/dev/null
[17:47:56.769018][cli][info] Found matching plugin vimeo for URL https://vimeo.com/720451389
[17:47:57.475643][plugins.vimeo][trace] {'url': 'https://live-api.vimeocdn.com/sessions/051dee4a-1ee1-47b6-8dad-65003ef74fb9/playlist/hls.m3u8?cdn=akamai-low-latency&~id=player&~exp=1655312400&~sig=0ySvobKv3d6vH_mSXPa_hEmMmIooNW68XjgqEAUI7rk'}
[17:47:57.915817][plugins.vimeo][trace] {'url': 'https://live-api.vimeocdn.com/sessions/051dee4a-1ee1-47b6-8dad-65003ef74fb9/playlist/hls.m3u8?cdn=fastly&~id=player&~exp=1655312400&~sig=VNOfp90Oc0XqnpuGyFOXBRV1h3HeOprk3hqap-KG8fA'}
[17:47:58.195956][cli][info] Available streams: 1080p_hls, 720p_hls, 540p_hls, 360p_hls, 240p_hls, 240p (worst), 360p, 540p, 720p, 1080p (best)
[17:47:58.202250][cli][info] Opening stream: 240p (hls)
[17:47:58.202751][cli][info] Writing output to
/dev/null
[17:47:58.202943][cli][debug] Checking file output
[17:47:58.204037][cli][debug] Pre-buffering 8192 bytes
[17:47:58.325602][stream.hls][debug] First Sequence: 6867; Last Sequence: 6872
[17:47:58.326063][stream.hls][debug] Start offset: 0; Duration: None; Start Sequence: 6870; End Sequence: None
[17:47:58.702279][cli][debug] Writing stream to output
[18:47:29.290146][plugins.vimeo][debug] Reloading Vimeo auth data
[18:47:31.027371][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[19:47:02.306376][plugins.vimeo][debug] Reloading Vimeo auth data
[19:47:04.267603][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[20:46:35.080602][plugins.vimeo][debug] Reloading Vimeo auth data
[20:46:36.895051][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[21:46:08.812545][plugins.vimeo][debug] Reloading Vimeo auth data
[21:46:10.496188][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[22:45:41.823225][plugins.vimeo][debug] Reloading Vimeo auth data
[22:45:42.941615][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[23:45:13.511936][plugins.vimeo][debug] Reloading Vimeo auth data
[23:45:14.970351][plugins.vimeo][debug] Reloaded Vimeo HLS URL
error: Error when reading from stream: Read timeout, exiting
[00:02:46.567462][cli][info] Stream ended
[00:02:46.570945][cli][info] Closing currently open stream...

real    374m51.699s
```
After muxed-HLS (filtered):
```
$ time sl/bin/streamlink -l trace -o /dev/null https://vimeo.com/720451389 240p_hls
[17:48:29.865706][cli][debug] OS:         Linux-4.9.0-18-amd64-x86_64-with-debian-9.13
[17:48:29.866365][cli][debug] Python:     3.7.3
[17:48:29.866624][cli][debug] Streamlink: 4.1.0+7.gc319aa44
[17:48:29.867917][cli][debug] Dependencies:
[17:48:29.872901][cli][debug]  isodate: 0.6.1
[17:48:29.874287][cli][debug]  lxml: 4.9.0
[17:48:29.875882][cli][debug]  pycountry: 22.3.5
[17:48:29.877390][cli][debug]  pycryptodome: 3.14.1
[17:48:29.878986][cli][debug]  PySocks: 1.7.1
[17:48:29.880088][cli][debug]  requests: 2.27.1
[17:48:29.881211][cli][debug]  websocket-client: 1.3.2
[17:48:29.882303][cli][debug]  importlib-metadata: 4.11.4
[17:48:29.883655][cli][debug] Arguments:
[17:48:29.883911][cli][debug]  url=https://vimeo.com/720451389
[17:48:29.884161][cli][debug]  stream=['240p_hls']
[17:48:29.884366][cli][debug]  --loglevel=trace
[17:48:29.884571][cli][debug]  --output=/dev/null
[17:48:29.885267][cli][info] Found matching plugin vimeo for URL https://vimeo.com/720451389
[17:48:30.424958][plugins.vimeo][trace] {'url': 'https://live-api.vimeocdn.com/sessions/051dee4a-1ee1-47b6-8dad-65003ef74fb9/playlist/hls.m3u8?cdn=akamai-low-latency&~id=player&~exp=1655312400&~sig=0ySvobKv3d6vH_mSXPa_hEmMmIooNW68XjgqEAUI7rk'}
[17:48:31.007946][plugins.vimeo][trace] {'url': 'https://live-api.vimeocdn.com/sessions/051dee4a-1ee1-47b6-8dad-65003ef74fb9/playlist/hls.m3u8?cdn=fastly&~id=player&~exp=1655312400&~sig=VNOfp90Oc0XqnpuGyFOXBRV1h3HeOprk3hqap-KG8fA'}
[17:48:31.266054][cli][info] Available streams: 1080p_hls, 720p_hls, 540p_hls, 360p_hls, 240p_hls, 240p (worst), 360p, 540p, 720p, 1080p (best)
[17:48:31.266610][cli][info] Opening stream: 240p_hls (hls-multi)
[17:48:31.267152][cli][info] Writing output to
/dev/null
[17:48:31.267404][cli][debug] Checking file output
[17:48:31.267761][stream.ffmpegmux][debug] Opening hls substream
[17:48:31.268930][stream.ffmpegmux][debug] Opening hls substream
[17:48:31.272222][utils.named_pipe][info] Creating pipe streamlinkpipe-6460-1-52
[17:48:31.275938][utils.named_pipe][info] Creating pipe streamlinkpipe-6460-2-1812
[17:48:31.277819][stream.ffmpegmux][debug] ffmpeg command: ffmpeg -nostats -y -i /tmp/streamlinkpipe-6460-1-52 -i /tmp/streamlinkpipe-6460-2-1812 -c:v copy -c:a copy -map 0:v? -map 0:a? -map 1:a -f mpegts pipe:1
[17:48:31.278557][stream.ffmpegmux][debug] Starting copy to pipe: /tmp/streamlinkpipe-6460-1-52
[17:48:31.279055][stream.ffmpegmux][debug] Starting copy to pipe: /tmp/streamlinkpipe-6460-2-1812
[17:48:31.296195][cli][debug] Pre-buffering 8192 bytes
[17:48:31.429234][stream.hls][debug] First Sequence: 6878; Last Sequence: 6883
[17:48:31.430883][stream.hls][debug] Start offset: 0; Duration: None; Start Sequence: 6881; End Sequence: None
[17:48:31.739175][stream.hls][debug] First Sequence: 6878; Last Sequence: 6883
[17:48:31.739353][stream.hls][debug] Start offset: 0; Duration: None; Start Sequence: 6881; End Sequence: None
[17:48:32.445282][cli][debug] Writing stream to output
[18:48:02.112475][plugins.vimeo][debug] Reloading Vimeo auth data
[18:48:02.434522][plugins.vimeo][debug] Waiting on Vimeo auth data update: 0s
[18:48:03.440440][plugins.vimeo][debug] Waiting on Vimeo auth data update: 1s
[18:48:03.702994][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[18:48:04.441782][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[19:47:34.289029][plugins.vimeo][debug] Reloading Vimeo auth data
[19:47:35.722887][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[19:47:35.746981][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[20:47:05.352472][plugins.vimeo][debug] Reloading Vimeo auth data
[20:47:06.346346][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[20:47:07.934939][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[21:46:06.405760][plugins.vimeo][debug] Reloading Vimeo auth data
[21:46:06.411638][plugins.vimeo][debug] Waiting on Vimeo auth data update: 0s
[21:46:07.412874][plugins.vimeo][debug] Waiting on Vimeo auth data update: 1s
[21:46:07.813955][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[21:46:08.414954][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[22:45:38.186605][plugins.vimeo][debug] Reloading Vimeo auth data
[22:45:38.634931][plugins.vimeo][debug] Waiting on Vimeo auth data update: 0s
[22:45:39.581262][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[22:45:39.636323][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[23:45:10.003283][plugins.vimeo][debug] Reloading Vimeo auth data
[23:45:11.304546][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[23:45:11.767777][plugins.vimeo][debug] Reloaded Vimeo HLS URL
[00:02:49.710648][stream.ffmpegmux][error] Pipe copy aborted: /tmp/streamlinkpipe-6460-2-1812
[00:02:49.725154][stream.ffmpegmux][debug] Pipe copy complete: /tmp/streamlinkpipe-6460-2-1812
[00:02:49.725601][stream.ffmpegmux][error] Pipe copy aborted: /tmp/streamlinkpipe-6460-1-52
[00:02:49.726809][stream.ffmpegmux][debug] Pipe copy complete: /tmp/streamlinkpipe-6460-1-52
[00:02:50.189536][stream.ffmpegmux][debug] Closing ffmpeg thread
[00:02:50.205087][stream.ffmpegmux][debug] Closed all the substreams
[00:02:50.206207][cli][info] Stream ended
[00:02:50.213892][cli][info] Closing currently open stream...

real    374m22.186s
```