@mkbloke 2022-06-16
If you have any advice on the Type checking failure, that would be great.  It is not something I am familiar with.  Thanks.