@mkbloke 2022-06-18
The latest push includes my patch from https://github.com/streamlink/streamlink/pull/4601#discussion_r899681531 added to @bastimeyer's stream.hls commit, just so the plugin can be tested for the time being.