@bastimeyer 2022-06-23
@mkbloke could you please post some vimeo URLs where the HLS playlist isn't `EXT-X-PLAYLIST-TYPE:VOD`? Otherwise it's an invariant playlist which contains every single segment, and it won't ever be reloaded by the worker, so the url property is only accessed once. The URLs you've posted in the OP are all of that type.

I've got some local changes with a couple of minor fixes, including plugin code you didn't touch in this PR. I don't want to add a comment for each of these changes. Take a look at the diff and squash with your commit unless you spot a mistake on my side. This doesn't contain any updates to the threading issues though.
https://github.com/streamlink/streamlink/compare/mkbloke:stream/hls/custom-muxed...bastimeyer:plugins/vimeo/4601