@mkbloke 2022-06-23
I've been getting live URLs from: https://vimeo.com/search?live=now&q=live

I've mostly been using the Phever live-streams, as they are available in both HLS and muxed-HLS.  They stream from about early afternoon to midnight or so (UK time, currently GMT+1): https://vimeo.com/phever