@mkbloke 2022-06-26
> This doesn't contain any updates to the threading issues though.

I assume you mean that the code as it is currently is not thread-safe?

I have looked at the threading stuff before and again today.  *I think* I have a solution that is thread-safe now.  I am testing it currently and will then push for you to take a look at if all goes well.